Imple/post/mapping member

+++ src/main/java/com/example/play/post/service/PostService.java
@@ -2,13 +2,17 @@ package com.example.play.post.service;
 
 import com.example.play.image.dto.ResponseImg;
 import com.example.play.image.service.PostImgService;
+import com.example.play.member.entity.Member;
+import com.example.play.member.service.MemberService;
 import com.example.play.post.constant.PageSize;
 import com.example.play.post.dto.RequestPostDto;
 import com.example.play.post.dto.ResponsePostOne;
 import com.example.play.post.dto.ResponsePostDTo;
 import com.example.play.post.dto.RequestUpdatePostDto;
 import com.example.play.post.entity.Post;
+import com.example.play.post.exception.PostDeleteException;
 import com.example.play.post.exception.PostNotFoundException;
+import com.example.play.post.exception.PostUpdateException;
 import com.example.play.post.postMapper.PostMapper;
 import com.example.play.post.repository.PostRepository;
 import com.example.play.post.repository.PostRepositoryCustom;
@@ -33,9 +37,11 @@ public class PostService {
     private final PostMapper postMapper;
     private final PostRepositoryCustom postRepositoryCustom;
     private final PostImgService postImgService;
+    private final MemberService memberService;
 
-    public ResponsePostOne create(RequestPostDto postDto, List<MultipartFile> files) {
-        Post post = postMapper.dtoToEntity(postDto);
+    public ResponsePostOne create(RequestPostDto postDto, List<MultipartFile> files, String email) {
+        Member member = memberService.findByEmail(email);
+        Post post = postMapper.dtoToEntity(postDto, member);
         Post saved = postRepository.save(post);
         if (!files.isEmpty()){
             List<ResponseImg> urlS = postImgService.savePostImage(files, saved);
@@ -68,8 +74,10 @@ public class PostService {
         return postMapper.pageEntityToDto(postPage);
     }
 
-    public ResponsePostOne update(Long postId , RequestUpdatePostDto updateDto, List<MultipartFile> files, List<Long> deleteImageList) {
+    public ResponsePostOne update(Long postId , RequestUpdatePostDto updateDto, List<MultipartFile> files,
+                                  List<Long> deleteImageList, String email) {
         Post post = findById(postId);
+        checkUpdateAuthorization(post, email);
         if (updateDto.getTitle() != null && !updateDto.getTitle().isEmpty()){
             post.changeTitle(updateDto.getTitle());
         }
@@ -79,11 +87,26 @@ public class PostService {
         List<ResponseImg> imgList = postImgService.update(post, deleteImageList, files);
         return postMapper.entityToDtoWithImage(post, imgList);
     }
-    public int delete(Long postId) {
+    private void checkUpdateAuthorization(Post post, String email){
+        Member member = memberService.findByEmail(email);
+        if (!post.getMember().equals(member)){
+            log.info("게시글을 작성하지 않은 멤버가 해당 게시글을 업데이트하고자 했습니다 게시글 아이디, 멤버 아이디 : {}", post.getId(), member.getId());
+            throw new PostUpdateException("유저가 해당 게시글을 업데이트할 수 있는 권한이 없습니다."+post.getId()+ member.getId());
+        }
+    }
+    public int delete(Long postId, String email) {
         Post post = findById(postId);
+        checkDeleteAuthorization(post, email);
         post.changeIsActive();
         postImgService.deleteImg(post);
         return post.getIsActive();
+    }
+    private void checkDeleteAuthorization(Post post, String email){
+        Member member = memberService.findByEmail(email);
+        if (!post.getMember().equals(member)){
+            log.info("게시글을 작성하지 않은 멤버가 해당 게시글을 삭제하고자 했습니다 게시글 아이디, 멤버 아이디 : {}", post.getId(), member.getId());
+            throw new PostDeleteException("유저가 해당 게시글을 삭제할 수 있는 권한이 없습니다."+post.getId()+ member.getId());
+        }
     }
     public ResponsePostOne entityToDto(Post post){
         List<ResponseImg> imgList = postImgService.readImages(post);
